requirements.txt
# Core ML frameworks
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.15.0
scikit-learn>=1.3.2

# NLP and transformers
transformers>=4.35.0
tokenizers>=0.15.0

# Data processing
pandas>=2.1.3
numpy>=1.24.3
Pillow>=10.1.0
albumentations>=1.3.0
datasets>=2.14.0
opencv-python>=4.8.0

# Visualization and tracking
matplotlib>=3.8.2
seaborn>=0.13.0
tensorboard>=2.15.0

# Optional but recommended
wandb>=0.16.0
tqdm>=4.66.1